Separate image processing functions into their own file
