+++ src/backend/vaapi/decoder.rs
@@ -37,9 +37,9 @@ use crate::decoder::DynHandle;
 use crate::decoder::FramePool;
 use crate::decoder::MappableHandle;
 use crate::decoder::StreamInfo;
-use crate::i4xx_copy;
-use crate::nv12_copy;
-use crate::y410_to_i410;
+use crate::image_processing::i4xx_copy;
+use crate::image_processing::nv12_copy;
+use crate::image_processing::y410_to_i410;
 use crate::DecodedFormat;
 use crate::Fourcc;
 use crate::Resolution;
